+++ CHANGELOG.md
@@ -5,3 +5,11 @@ All notable changes to this project will be documented in this file.
 ## v2.0.0
 
 Initial Release
+
+## v2.0.1
+
+* Add support for centering map by first position
+* Add support for andpath view with markers
+
+## v2.0.2
+* Add support for tooltips from data